pyproject.toml
[project]
name = "keenetica"
version = "0.1.0"
description = ""
authors = [
    {name = "vladimir",email = "vladimir@digital-order.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "pyqt6 (>=6.9.0,<7.0.0)",
    "pytest (>=8.3.5,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
